Add ESLint and Prettier to encourage code hygiene
ESLint is is a linter that checks the code for potential problems in the code,
such as duplicate variables or unsafe use of certain functions. It's a safety
net around common mistakes. With this change, it can be run from the command
line using `yarn lint`, though installing the eslint plugin is advisable.

Prettier is a code formatter. It is best run "on save" using the prettier code
editor plugin to automatically snap the code into a consistent formatter. Again,
it's best if it's something you don't have to think about.

I've gone ahead and enable eslint warnings for prettier violations, since that
makes it more obvious while working on code. But if that gets annoying we can
always disable it.

Along with this, I moved the main project folders into a `src` directory to keep
them together and allow for the addition of folders without needing to keep
a running list in the config.