Set up GitHub action to run ESLint against PRs
This is a handy way to avoid forgetting to lint/format. The ESLint command runs
against every commit pushed to the repo.

This also adds version files, specifying NodeJS 12.x, the current LTS release,
at the root of the project. It was also added to "engines" in the `package.json`
which is what Heroku uses to determine the NodeJS version.