Add Course Detail Page
Adds a course detail page accessible from the Course List page. The course is
fetched server-side so that the title can be correctly populated. Then the
course instances are fetched async.

The server-side fetching requires some additional logic. If the user is not
authenticated, we return early with empty props. That way the database is not
queried, potentially leaking sensitive details. We then handle the client-side
unauthorized redirect using the protect page hook.